refactor(TextField): remove the interface to index file
