+++ src/components/inputs/TextField/index.tsx
@@ -1,9 +1,36 @@
 import { useState } from "react";
-import { ITextFieldProps } from "./interfaces/TextField.interface";
+
 import { InputType, inputTypes } from "./types/TextField.InputType.type";
 import { State, states } from "./types/TextField.State.type";
-
+import { Size } from "./types/TextField.Size.type";
 import { TextFieldUI } from "./interface";
+
+export interface ITextFieldProps {
+  label?: string;
+  name: string;
+  id: string;
+  placeholder: string;
+  isDisabled?: boolean;
+  type?: InputType;
+  value?: string | number;
+  handleChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  iconBefore?: React.ReactNode;
+  iconAfter?: React.ReactNode;
+  maxLength?: number;
+  minLength?: number;
+  max?: number;
+  min?: number;
+  isRequired: boolean;
+  state?: State;
+  errorMessage?: string;
+  validMessage?: string;
+  size?: Size;
+  isFullWidth?: boolean;
+  handleFocus?: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  handleBlur?: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  readOnly?: boolean;
+  isFocused?: boolean;
+}
 
 const defaultIsDisabled = false;
 const defaultType: InputType = "text";
