+++ src/components/inputs/TextField/interface.tsx
@@ -1,10 +1,11 @@
-import { ITextFieldProps } from "./interfaces/TextField.interface";
-import { IMessageProps } from "./interfaces/TextField.Message.interface";
-
 import { MdOutlineError, MdCheckCircle } from "react-icons/md";
 
 import { Label } from "../Label";
 import { Text } from "../../data/Text";
+
+import { Size } from "./types/TextField.Size.type";
+import { State } from "./types/TextField.State.type";
+import { ITextFieldProps } from ".";
 
 import {
   StyledContainer,
@@ -15,7 +16,13 @@ import {
   StyledErrorMessageContainer,
   StyledValidMessageContainer,
 } from "./styles";
-import { Size } from "./types/TextField.Size.type";
+
+export interface IMessageProps {
+  state?: State;
+  isDisabled?: boolean;
+  errorMessage?: string;
+  validMessage?: string;
+}
 
 const getTypo = (size: Size) => {
   if (size === "compact") {
