--- src/components/inputs/TextField/interfaces/TextField.Message.interface.ts
@@ -1,8 +1,0 @@
-import { State } from "../types/TextField.State.type";
-
-export interface IMessageProps {
-  state?: State;
-  isDisabled?: boolean;
-  errorMessage?: string;
-  validMessage?: string;
-}
